all transaction, deposit and withdrawal

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->enum('account_type', ['Individual', 'Business']);
-            $table->decimal('balance')->default(0);
+            $table->double('balance')->default(0);
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
